Add files via upload
The bitstream_in file contains a pre-generated packet, consisting of the CCSDS preamble followed by the payload data. This packet must be transmitted by the transmitter in repetition mode and provided to GNU Radio as a file source. This allows the received signal to be correctly aligned with the transmitted one and enables the computation of the Bit Error Rate (BER).